Moved SuicidalAOEAttackBattleEffect to battle directory.
