+++ app/src/main/java/noteworthyengine/units/PufferZug.java
@@ -10,6 +10,7 @@ import noteworthyengine.RenderNode;
 import noteworthyengine.RenderSystem;
 import noteworthyengine.SeparationNode;
 import noteworthyengine.battle.BattleNode;
+import noteworthyengine.battle.SuicidalAOEAttackBattleEffect;
 import noteworthyengine.players.PlayerUnit;
 import noteworthyframework.Unit;
 import structure.RewriteOnlyArray;
